# lib/api-zod/src/generated/types/visitInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VisitInput {
  /** Name of the page visited (e.g. "ask", "yes", "boba") */
  page: string;
  /** Number of seconds spent on the page */
  timeSpentSeconds: number;
  /**
     * User agent string
     * @nullable
     */
  userAgent?: string | null;
  /**
     * How many times the user clicked No (for the ask page)
     * @nullable
     */
  noCount?: number | null;
  /**
     * The final answer given (yes / no / boba)
     * @nullable
     */
  finalAnswer?: string | null;
}