# lib/api-zod/src/generated/types/visitLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VisitLog {
  id: number;
  page: string;
  timeSpentSeconds: number;
  ipAddress: string;
  /** @nullable */
  userAgent?: string | null;
  /** @nullable */
  noCount?: number | null;
  /** @nullable */
  finalAnswer?: string | null;
  visitedAt: Date;
}